+++ Assets/DeltaDNA/Ads/Editor/Networks/IosNetworks.cs
@@ -22,7 +22,7 @@ namespace DeltaDNA.Ads.Editor {
     
     internal sealed class IosNetworks : Networks {
         
-        private const string VERSION = "~> 1.8.0";
+        private const string VERSION = "~> 1.9.0";
         private const string VERSION_DEBUG = "~> 1.0.0";
         
         private readonly object[] sources = new object[] {